Add L1GasCost (Please rework)
This is extremely invasive, and, I feel like there must be a better way.
But, this was the approach taken by op-geth, so, trying it for now.

Notably, I've left out the staged sync exec3 path, which... may be
catastrophic, but it wasn't at all obvious how to expose the evm state
at the time the block context is constructed.  But, it seems to work in
the optimism e2e, so, proceeding with this for now.