Add L1Fee fields to receipt
Note, this only implements the "Fee" and not the other fields.  We need
to adjust wiring generally for L1Cost

Also, I've very hackily disabled the caching path because we would have
to either change the db format to store more, or, more likely wire the
block state into reconstructing the receipt.