Disable short-circuit on setting block head hash
When an Optimism chain re-orgs, it can do so to a previous block in the
canonical chain.  Ethereum does not allow this behavior, and, presumably
the Erigon code is assuming that a new block must commit and advance the
block head hash.  So, prior to this commit, Erigon would not set hashes
if the block number of the headed head was less than or equal execution
block.