Do not re-org when canonical is HEAD
Traditional Ethereum ignores forkchoice requests which are to a point in
the canonical chain.  Optimism needs to allow 'arbitrarily deep re-orgs'
including into the canonical chain.  So, we added a check to explicitly
disallow this 'ignore'.  But, this means we re-org every fork choice,
even when the forkchoice is to the current canonical head.  This change
allows the forkchoice to be ignored when the forkchoice is truly a
no-op.