Fixes from scrutinizing a diff against upstream
The integration tests were failing based on something that changed
during the latest merge.  Because of this, I decided to audit what was
different between our fork and the upstream project.  This set of
changes includes small changes to more closely align us with the
upstream code, especially in inconsequential areas of formatting.

Additionally, this change fixes the real bug which was breaking the
integration tests -- that the blob txes were being accidentally being
encoded twice.